# api/docs/swagger.yaml
definitions:
...
  models.CreateGroup:
    properties:
      branch_id:
        type: string
      group_id:
        type: string
      teacher_id:
        type: string
      type:
        type: string
    type: object
  models.CreateLesson:
    properties:
      from:
        type: string
      group_id:
        type: string
      schedule_id:
        type: string
      theme:
        type: string
      to:
        type: string
    type: object
...
  models.GetAllGroupsResponse:
    properties:
      count:
        type: integer
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
    type: object
  models.GetAllLessonsResponse:
    properties:
      count:
        type: integer
      lessons:
        items:
          $ref: '#/definitions/models.Lesson'
        type: array
    type: object
...
  models.GetGroup:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      group_id:
        type: string
      id:
        type: string
      teacher_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.GetLesson:
    properties:
      created_at:
        type: string
      from:
        type: string
      group_id:
        type: string
      id:
        type: string
      schedule_id:
        type: string
      theme:
        type: string
      to:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Group:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      group_id:
        type: string
      id:
        type: string
      teacher_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.Lesson:
    properties:
      created_at:
        type: string
      from:
        type: string
      group_id:
        type: string
      id:
        type: string
      schedule_id:
        type: string
      theme:
        type: string
      to:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.UpdateGroup:
    properties:
      branch_id:
        type: string
      group_id:
        type: string
      teacher_id:
        type: string
      type:
        type: string
    type: object
  models.UpdateLesson:
    properties:
      from:
        type: string
      group_id:
        type: string
      schedule_id:
        type: string
      theme:
        type: string
      to:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /group:
    get:
      description: This API returns group list
      parameters:
      - description: page number
        in: query
        name: page
        type: integer
      - description: limit per page
        in: query
        name: limit
        type: integer
      - description: search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllGroupsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: get all Group
      tags:
      - group
    post:
      consumes:
      - application/json
      description: This api is creates a new group and returns its id
      parameters:
      - description: group
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.CreateGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: create a group
      tags:
      - group
  /group/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a group by its ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: delete a group by ID
      tags:
      - group
    get:
      consumes:
      - application/json
      description: Retrieves a group by its ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: return a group by ID
      tags:
      - group
    put:
      consumes:
      - application/json
      parameters:
      - description: group ID
        in: path
        name: id
        required: true
        type: string
      - description: group
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.UpdateGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: update a group
      tags:
      - group
  /lesson:
    get:
      description: This API returns lesson list
      parameters:
      - description: page number
        in: query
        name: page
        type: integer
      - description: limit per page
        in: query
        name: limit
        type: integer
      - description: search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllLessonsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get All Lessons
      tags:
      - lesson
    post:
      consumes:
      - application/json
      description: This api is creates a new lesson and returns its id
      parameters:
      - description: lesson
        in: body
        name: lesson
        required: true
        schema:
          $ref: '#/definitions/models.CreateLesson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Lesson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: create a lesson
      tags:
      - lesson
  /lesson/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a lesson by its ID
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Deletes a lesson by ID
      tags:
      - lesson
    get:
      consumes:
      - application/json
      description: Retrieves a lesson by its ID
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetLesson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: return a lesson by ID
      tags:
      - lesson
    put:
      consumes:
      - application/json
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: string
      - description: Lesson
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/models.UpdateLesson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Lesson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: update a lesson
      tags:
      - lesson
...
swagger: "2.0"